requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
captcha==0.4
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
mailtrap==2.0.1
passlib==1.7.4
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
